% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventid2name.R
\docType{data}
\name{eventid2name}
\alias{eventid2name}
\title{Event ID to Name mapping}
\format{
Data frame with columns
\describe{
\item{event}{Event ID}
\item{subevent}{Subevent ID}
\item{event_label}{Event description}
\item{subevent_label}{Subevent description}
}
}
\source{
https://doi.org/10.6084/m9.figshare.c.4415000.v5
}
\usage{
eventid2name
}
\description{
Mapping table of numeric Event ID to Event Name.
}
\examples{
  dim(eventid2name)
  head(eventid2name)
}
\keyword{datasets}
