% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socceR.R
\docType{data}
\name{fifa2018}
\alias{fifa2018}
\title{FIFA 2018 prediction matrices}
\format{A list with 5 predictions (each a 7 by 32 matrix) containing the predictions probabilities of 1st, 2nd, 3rd, 4th, 5th-8th, 9th-12th, and 17th-32nd place.
\describe{
  \item{flat}{A prediction with equal probability of winning for all teams}
  \item{ekstrom1}{Ekstrom's prediction (based on the Skellam distribution)}
  \item{ekstrom2}{Ekstrom's prediction (based on the ELO rankings)}
  \item{GLSE1}{Prediction of Groll et all}
  \item{GLSE2}{Updated prediction of Groll et all}
}}
\source{
\url{http://sandsynligvis.dk/2018/08/03/world-cup-prediction-winners/}
}
\usage{
fifa2018
}
\description{
A list containing five predictions for the FIFA 2018 World Cup.
}
\keyword{datasets}
