% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socceR.R
\docType{data}
\name{fifa2018result}
\alias{fifa2018result}
\title{FIFA 2018 end results}
\format{A vector of the final rankings}
\source{
\url{http://sandsynligvis.dk/2018/08/03/world-cup-prediction-winners/}
}
\usage{
fifa2018result
}
\description{
A named vector sorted in the ranking of the teams in the FIFA 2018 World Cup. The value correspond to the corresponding columns in the prediction matrices of fifa2018
}
\keyword{datasets}
