% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rps}
\alias{rps}
\title{Computes the rank probability score for a tournament}
\usage{
rps(m, outcome, rankweights = 1L)
}
\arguments{
\item{m}{A prediction matrix}

\item{outcome}{A prediction matrix}

\item{rankweights}{A prediction matrix}
}
\value{
A vector or list with the NA's replaced by the last observed value.
}
\description{
Fill down missing values with the latest non-missing value
}
\examples{


}
\author{
Claus Ekstrom <ekstrom@sund.ku.dk>
}
