% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socceR.R
\docType{package}
\name{socceR}
\alias{socceR}
\alias{socceR-package}
\title{Evaluating sport tournament predictions}
\description{
Functions for evaluating sport tournament predictions, the tournament rank probability score, and working with models for prediction sport matches.
}
\author{
Claus Ekstrom <ekstrom@sund.ku.dk>
}
