% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags2name.R
\docType{data}
\name{tags2name}
\alias{tags2name}
\title{Tags to Name mapping}
\format{
Data frame with columns
\describe{
\item{Tag}{Tag ID}
\item{Label}{Event label}
\item{Description}{Event description}
}
}
\source{
https://doi.org/10.6084/m9.figshare.c.4415000.v5
}
\usage{
tags2name
}
\description{
Mapping table of numeric Tag to Event Name.
}
\examples{
  dim(tags2name)
  head(tags2name)
}
\keyword{datasets}
