% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_prediction.R
\name{collapse_prediction}
\alias{collapse_prediction}
\title{Convert factor to numeric vector}
\usage{
collapse_prediction(ranks = c(1, 2, 3, 4, 8, 16, 32))
}
\arguments{
\item{ranks}{A factor}
}
\value{
Returns a numeric vector of the same length as x
}
\description{
Converts the factor labels to numeric values and returns the factor as a numeric vector
}
\details{
Returns a vector of numeric values. Elements in the input factor that cannot be converted to numeric will produce NA.
}
\examples{


}
\author{
Claus Ekstrom \email{ekstrom@sund.ku.dk}
}
\keyword{manip}
