% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_weights.R
\name{optimize_weights}
\alias{optimize_weights}
\title{Optimize weights from list of prediction matrices}
\usage{
optimize_weights(predictionlist, outcome, FUN = rps)
}
\arguments{
\item{predictionlist}{A factor}

\item{outcome}{An integer vector listing the}

\item{FUN}{}
}
\value{
Returns a numeric vector containing an optimal of the weights of the
}
\description{
Converts the factor labels to numeric values and returns the factor as a numeric vector
}
\details{
Returns a vector of numeric values. Elements in the input factor that cannot be converted to numeric will produce NA.
}
\examples{

m1 <- matrix(c(1, 0, 0, 0, 0, 1, 0, 0, 0, 0, .5, .5, 0, 0, .5, .5), 4)
m1 # Prediction where certain on the top ranks
m2 <- matrix(c(.5, .5, 0, 0, .5, .5, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1), 4)
m2  # Prediction where the groups are okay 
m3 <- matrix(c(.5, .5, 0, 0, .5, .5, 0, 0, 0, 0, .5, .5, 0, 0, .5, .5), 4)
m3  # Prediction where no clue about anything
m4 <- matrix(rep(1/4, 16), 4)

optimize_weights(list(m1, m2, m3, m4), 1:4)

}
\author{
Claus Ekstrom \email{claus@rprimer.dk}
}
\keyword{manip}
